requirements.txt
streamlit
playwright
playwright-stealth
google-generativeai